Add widget rendering to sidebar with REST API.
To allow for the flexibility of widgets, I've takedn the widget controller to expose the data via the REST API, coupled with the react rendering. Specific widgets are supported, and future ones would need to be manually added to the map.

Couple of screenshots:

Text widget and unknown pages widget:

![](https://joehoyle-captured.s3.amazonaws.com/qK1mFqAl.png)