Explicitly specify "h2" as script handle
This project will not load properly if the repo is cloned as-is (with
the capitalized "H2") because it reads the script handle from the theme
directory's name. Since we manually use "h2" as the handle for script
localization, specifying that same identifier manually ensures the theme
will run wherever it is deployed.