Fix issue where users may show up twice for the same reaction
This is probably caused by submitting an emoji twice before it
had registered, but users can sometimes be listed twice for the
same emoji. This is not how reacts were designed to work, and
it triggers a noisy non-unique key warning in the console.
Running .includes on each array does incur a little extra looping
while rendering, but emoji lists are never so long this should
be a noticeable impact on the page render time.