Prevent SVG loading indicators from expanding page width
Fixes #387

I tracked this down by pausing JS script execution while the page was
loading in a narrow viewport, which let me see that the loading elements
were displaying at a wide fixed width that exceeded the dimensions of
the viewport. Wrapping mobile loaders in a div that cannot exceed the
page dimensions appears to fix the issue in my browser; the width can
be adjusted if, after deploy, it still doesn't look right on mobile.